fundo_project/requirements.txt
asgiref==3.2.7
astroid==2.4.0
atomicwrites==1.4.0
attrs==19.3.0
boto3==1.14.23
botocore==1.17.23
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
CodeConvert==3.0.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.5
django-admin==2.0.1
django-colorfield==0.3.2
django-colorful==1.3
<<<<<<< HEAD
=======
django-elasticsearch-dsl==7.1.4
>>>>>>> services_and_search
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-logpipe==1.0.0
django-models-ext==1.1.9
django-redis-cache==2.1.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-short-url==1.1.4
django-shorturls==2.0.0
django-six==1.0.4
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
Faker==0.9.1
fixture==1.5.11
furl==2.1.0
idna==2.9
isodate==0.6.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
json-tricks==3.15.2
jwt==1.0.0
kafka-python==2.0.1
lazy-object-proxy==1.4.3
lru-dict==1.1.6
mark==0.2
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
mixture==0.2.1
mock==4.0.2
more-itertools==8.2.0
oauthlib==3.1.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==20.3
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.0
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
pytest-mock==3.2.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.1.0
pytz==2020.1
rdflib==5.0.0
redis==3.5.1
requests==2.23.0
requests-aws4auth==1.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
screen==1.0.1
shortuuid==1.0.1
simplejson==3.17.2
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
text-unidecode==1.2
TimeConvert==1.5.1
toml==0.10.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
xlwt==1.3.0
